# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gymtime API",
    "description": "Gymtime health and invitation-only sign-in workflows.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "tags": [
          "accounts"
        ],
        "summary": "# Errors\nReturns safe authentication, organizer permission, or storage errors.",
        "operationId": "list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponse"
                  }
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "# Errors\nReturns safe authentication, input, organizer permission, or storage errors.",
        "operationId": "invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{id}": {
      "patch": {
        "tags": [
          "accounts"
        ],
        "summary": "# Errors\nReturns safe authentication, input, last-organizer, permission, or storage errors.",
        "operationId": "status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {},
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/request-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe input, origin, rate limit, or infrastructure errors.",
        "operationId": "request_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeResponse"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe session or infrastructure errors.",
        "operationId": "session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "# Errors\nReturns safe input, origin, code, or infrastructure errors.",
        "operationId": "verify_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyCodeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "organizer",
          "enabled"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "organizer": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CodeResponse": {
        "type": "object",
        "required": [
          "message",
          "resend_after_seconds"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "resend_after_seconds": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id",
          "issues"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldIssue"
            }
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
      "FieldIssue": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "InvitationRole": {
        "type": "string",
        "enum": [
          "organizer",
          "coach"
        ]
      },
      "InviteBody": {
        "type": "object",
        "required": [
          "email",
          "role"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/InvitationRole"
          }
        },
        "additionalProperties": false
      },
...
      "RequestCodeBody": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "user_id",
          "email",
          "organizer",
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "organizer": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "StatusBody": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "VerifyCodeBody": {
        "type": "object",
        "required": [
          "email",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  }
}